Partial support for -08. Varint encodings, some frame changes, etc.

+++ lib/include/quant/quant.h
@@ -74,7 +74,7 @@ extern void __attribute__((nonnull)) q_free(struct w_iov_sq * const q);
 
 extern uint64_t __attribute__((nonnull)) q_cid(const struct q_conn * const c);
 
-extern uint32_t __attribute__((nonnull)) q_sid(const struct q_stream * const s);
+extern uint64_t __attribute__((nonnull)) q_sid(const struct q_stream * const s);
 
 extern void __attribute__((nonnull))
 q_write_str(void * const q, struct q_stream * const s, const char * const str);
